backend/pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Multi-agent orchestration platform with MCP support"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    
    # MCP SDK
    "mcp>=1.0.0",
    
    # LLM orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "litellm>=1.50.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    
    # Async utilities
    "httpx>=0.27.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
